% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{App}
\usage{
app(spatial, df, interventions = c("itn", "irs"), spatial_id = "NAME_1")
}
\arguments{
\item{spatial}{The sf spatial data}

\item{df}{The simulation bank}

\item{interventions}{Vector of intervention options}

\item{spatial_id}{Name of unique spatial unit identifier column}
}
\description{
Launch shiny app
}
