% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{apply_coverage}
\alias{apply_coverage}
\title{Multiply the boolean choice matrix by the coverage vector}
\usage{
apply_coverage(rv, coverage)
}
\arguments{
\item{rv}{Reactive coverage vector}

\item{coverage}{Reactive choice matrix}
}
\description{
Multiply the boolean choice matrix by the coverage vector
}
