% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{available}
\alias{available}
\title{Pull out names of spatial units where an intervention is available}
\usage{
available(mat, intervention)
}
\arguments{
\item{mat}{choice matrix}

\item{intervention}{Intervention name}
}
\description{
Pull out names of spatial units where an intervention is available
}
