% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_input_df}
\alias{check_input_df}
\title{Column checks input data.frame}
\usage{
check_input_df(input_df, spatial_id)
}
\arguments{
\item{input_df}{Input data.frame}

\item{spatial_id}{Name of unique spatial unit identifier column}
}
\description{
Column checks input data.frame
}
