% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_input_spatial}
\alias{check_input_spatial}
\title{Column checks spatial input}
\usage{
check_input_spatial(input_spatial, spatial_id)
}
\arguments{
\item{input_spatial}{Input spatial data}

\item{spatial_id}{Name of unique spatial unit identifier column}
}
\description{
Column checks spatial input
}
