% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{clear}
\alias{clear}
\title{Clear the intervention matrix for selected intervention and spatial unit
combinations}
\usage{
clear(mat, intervention, units = NULL)
}
\arguments{
\item{mat}{choice matix}

\item{intervention}{Intervention name}

\item{units}{Vector of spatial unit names}
}
\description{
Clear the intervention matrix for selected intervention and spatial unit
combinations
}
