% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_mwi}
\alias{df_mwi}
\title{Output data MWI}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112 rows and 11 columns.
}
\usage{
df_mwi
}
\description{
A data.frame
}
\keyword{datasets}
