% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{df_pd}
\alias{df_pd}
\title{Select current solution from simulation bank}
\usage{
df_pd(df, rv, interventions, spatial_id, bau_impact, bau_cost)
}
\arguments{
\item{df}{Simulation bank}

\item{rv}{Reactive values}

\item{interventions}{Vector of intervention options}

\item{spatial_id}{Name of unique spatial unit identifier column}

\item{bau_impact}{BAU impact}

\item{bau_cost}{BAU cost}
}
\value{
Impact plot data
}
\description{
Select current solution from simulation bank
}
