% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_sdn2}
\alias{df_sdn2}
\title{Output data sdn}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 320 rows and 13 columns.
}
\usage{
df_sdn2
}
\description{
A data.frame
}
\keyword{datasets}
