% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{df_selection}
\alias{df_selection}
\title{Select current solution from simulation bank}
\usage{
df_selection(df, selected_itn, selected_irs)
}
\arguments{
\item{df}{Simulation bank}

\item{selected_itn}{Current ITN sub unit selection}

\item{selected_irs}{Current IRS sub unit selection}
}
\value{
Filtered simulation bank
}
\description{
Select current solution from simulation bank
}
