% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_bau_impact}
\alias{get_bau_impact}
\title{Current impact (BAU)}
\usage{
get_bau_impact(df)
}
\arguments{
\item{df}{Input data}
}
\value{
Vector of maximum cases and deaths averted
}
\description{
Current impact (BAU)
}
