% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_module.R
\name{impactServer}
\alias{impactServer}
\title{Impact module UI}
\usage{
impactServer(id, rv, coverage, df)
}
\arguments{
\item{id}{Impact ID}

\item{rv}{Reactive values}

\item{coverage}{coverage reactive values}

\item{df}{The simulation bank}
}
\description{
Impact module UI
}
