% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{implemented}
\alias{implemented}
\title{Pull out names of spatial units where an intervention is implemented}
\usage{
implemented(mat, intervention)
}
\arguments{
\item{mat}{choice matix}

\item{intervention}{Intervention name}
}
\description{
Pull out names of spatial units where an intervention is implemented
}
