% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{init_matrix}
\alias{init_matrix}
\title{Initialise a spatial unit x interventions boolean matrix}
\usage{
init_matrix(units, interventions)
}
\arguments{
\item{units}{Vector of spatial unit names}

\item{interventions}{Vector of intervention names}
}
\value{
Boolean matrix
}
\description{
Initialise a spatial unit x interventions boolean matrix
}
