% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{intervention_mix_map}
\alias{intervention_mix_map}
\title{Create the intervention mix map}
\usage{
intervention_mix_map(spatial, bbox, mix_options, cols)
}
\arguments{
\item{spatial}{The sf spatial data}

\item{bbox}{Bounding box for map}

\item{mix_options}{Vector of possible intervention mixes}

\item{cols}{Vector of mix option colours}
}
\description{
Create the intervention mix map
}
