% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_cols}
\alias{map_cols}
\title{Draw map colours}
\usage{
map_cols(interventions)
}
\arguments{
\item{interventions}{Vector of interventions}
}
\value{
A colour vector
}
\description{
Each of up to a maximum of 8 interventions is assigned a named colour
}
