% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_module.R
\name{mapServer}
\alias{mapServer}
\title{Map module server}
\usage{
mapServer(
  id,
  rv,
  coverage,
  current_matrix,
  strata_matrix,
  col,
  spatial,
  bbox,
  n_strata,
  session
)
}
\arguments{
\item{id}{Intervention ID}

\item{rv}{Reactive values}

\item{coverage}{coverage reactive values}

\item{current_matrix}{Current implemented areas}

\item{strata_matrix}{Strata matrices}

\item{col}{Intervention colour}

\item{spatial}{Spatial data for intervention}

\item{bbox}{Map bounding box}

\item{n_strata}{Number of stratification levels}

\item{session}{app session}
}
\description{
Map module server
}
