% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_module.R
\name{mapUI}
\alias{mapUI}
\title{Map module UI}
\usage{
mapUI(id, n_strata, admin_units, coverage_choices, coverage)
}
\arguments{
\item{id}{Intervention ID}

\item{n_strata}{Number of stratification levels}

\item{admin_units}{spatial spatial unit names}

\item{coverage_choices}{Available coverage choices}

\item{coverage}{Reactive coverage selection}
}
\description{
Map module UI
}
