% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{mat}
\alias{mat}
\title{Create matrix for optimisation}
\usage{
mat(unit, y)
}
\arguments{
\item{unit}{Sub unit name}

\item{y}{Matrix input}
}
\value{
Matrix in format for optimisation using the om package
}
\description{
Create matrix for optimisation
}
