% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwi}
\alias{mwi}
\title{Spatial data}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 28 rows and 12 columns.
}
\usage{
mwi
}
\description{
A simple feature spatial data frame
}
\keyword{datasets}
