% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optimise}
\alias{optimise}
\title{Run optimisation}
\usage{
optimise(target, budget)
}
\arguments{
\item{target}{Target for optimisation, can be "cases_averted" or "deaths_averted".}

\item{budget}{Budget}
}
\description{
Run optimisation
}
