% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{overlap_map}
\alias{overlap_map}
\title{Render a polygon fill overlay}
\usage{
overlap_map(proxymap, data, colour)
}
\arguments{
\item{proxymap}{Proxy map}

\item{data}{sf data}

\item{colour}{Fill colour}
}
\description{
Current fixed to render NAME_1 level
}
