% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Time series plot plot}
\usage{
plot_time_series(pd, maxy)
}
\arguments{
\item{pd}{plot data}

\item{maxy}{max y value}
}
\description{
Time series plot plot
}
