% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{reverse}
\alias{reverse}
\title{Reverse the intervention matrix for selected intervention and spatial unit
combinations}
\usage{
reverse(mat, intervention, units = NULL)
}
\arguments{
\item{mat}{choice matrix}

\item{intervention}{Intervention name}

\item{units}{Vector of spatial unit names}
}
\description{
Reverse the intervention matrix for selected intervention and spatial unit
combinations
}
