% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{select}
\alias{select}
\title{Select the intervention matrix for selected intervention and spatial unit
combinations}
\usage{
select(mat, intervention, units = NULL)
}
\arguments{
\item{mat}{choice matix}

\item{intervention}{Intervention name}

\item{units}{Vector of spatial unit names}
}
\description{
Select the intervention matrix for selected intervention and spatial unit
combinations
}
