% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratification_colours}
\alias{stratification_colours}
\title{Stratification colour palette}
\format{
An object of class \code{character} of length 4.
}
\usage{
stratification_colours
}
\description{
A vector
}
\keyword{datasets}
