% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Time series plot plot}
\usage{
time_series_plot(pd, maxy)
}
\arguments{
\item{pd}{plot data}

\item{maxy}{max y value}
}
\description{
Time series plot plot
}
