% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{update_intervention_mix_map}
\alias{update_intervention_mix_map}
\title{Update the intervention mix map}
\usage{
update_intervention_mix_map(
  selection,
  spatial,
  interventions,
  bbox,
  mix_options,
  cols
)
}
\arguments{
\item{selection}{The reactive choice matrix}

\item{spatial}{The sf spatial data}

\item{interventions}{Interventions vector}

\item{bbox}{Bounding box for map}

\item{mix_options}{Vector of possible intervention mixes}

\item{cols}{Vector of mix option colours}
}
\description{
Update the intervention mix map
}
