% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{df_pd}
\alias{df_pd}
\title{Select current solution from simulation bank}
\usage{
df_pd(df, rv, interventions)
}
\arguments{
\item{df}{Simulation bank}

\item{rv}{Reactive values}

\item{interventions}{Vector of intervention options}
}
\value{
Impact plot data
}
\description{
Select current solution from simulation bank
}
