% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_max_impact}
\alias{get_max_impact}
\title{Maximum impact possible}
\usage{
get_max_impact(df)
}
\arguments{
\item{df}{Input data}
}
\value{
Vector of maximum cases and deaths averted
}
\description{
Maximum impact possible
}
