% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_module.R
\name{mapServer}
\alias{mapServer}
\title{Map module server}
\usage{
mapServer(
  id,
  rv,
  all,
  current,
  col,
  rankings,
  spatial,
  spatial_id,
  n_strata,
  strata_selection
)
}
\arguments{
\item{id}{Intervention ID}

\item{rv}{Reactive values}

\item{all}{All subunits}

\item{current}{Current subunits}

\item{col}{Intervention colour}

\item{rankings}{CE ranking table}

\item{spatial}{The sf spatial data}

\item{spatial_id}{Name of unique spatial unit identifier column}

\item{n_strata}{Number of stratification levels}

\item{strata_selection}{List of selected subunits corresponding to strata selection buttons}
}
\description{
Map module server
}
