% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_module.R
\name{mapServer}
\alias{mapServer}
\title{Map module server}
\usage{
mapServer(id, rv, all, current, col)
}
\arguments{
\item{id}{Intervention ID}

\item{all}{All subunits}

\item{current}{Current subunits}

\item{overwrite}{Plotting input}

\item{trigger}{Trigger for map update}
}
\description{
Map module server
}
