% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_module.R
\name{mapUI}
\alias{mapUI}
\title{Map module UI}
\usage{
mapUI(id, n_strata, admin_units)
}
\arguments{
\item{id}{Intervention ID}

\item{n_strata}{Number of stratification levels}

\item{admin_units}{Available spatial unit names}
}
\description{
Map module UI
}
