% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdn2}
\alias{sdn2}
\title{Spatial data SDN admin 2}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 189 rows and 6 columns.
}
\usage{
sdn2
}
\description{
A simple feature spatial data frame
}
\keyword{datasets}
