% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{stratification_map}
\alias{stratification_map}
\title{Create strata map}
\usage{
stratification_map(spatial, df, n_strata, bbox)
}
\arguments{
\item{spatial}{The sf spatial data}

\item{df}{The simulation bank}

\item{n_strata}{Number of strata}

\item{bbox}{Bounding box for map}
}
\description{
Create strata map
}
